+++ test/unit/test_spec.py
@@ -1,6 +1,9 @@
+from unittest.mock import Mock
 import yaml
 
-from k8s_jobs.spec import YamlFileSpecSource
+from kubernetes.client import V1Job, V1ObjectMeta
+
+from k8s_jobs.spec import JobGenerator, StaticJobSpecSource, YamlFileSpecSource
 
 
 class TestSpecSource:
@@ -27,3 +30,40 @@ class TestSpecSource:
         c = YamlFileSpecSource(str(tmp_file_name))
 
         assert {"biz": "foo"} == c.get(template_args={"buzz": "foo"})
+
+
+class TestJobGenerator:
+    def test_unique_names(self):
+        generator = JobGenerator(
+            StaticJobSpecSource(
+                V1Job(metadata=V1ObjectMeta(name="iloveyouabushelandapeck"))
+            )
+        )
+
+        j1 = generator.generate()
+        j2 = generator.generate()
+
+        assert (
+            j1.metadata.name != j2.metadata.name
+        ), "Each generated job must have a unique name"
+
+    def test_generate_with_dict_config(self):
+        job = V1Job(metadata=V1ObjectMeta(name="iloveyouabushelandapeck"))
+        generator = JobGenerator(StaticJobSpecSource(job.to_dict()))
+
+        j = generator.generate()
+        assert (
+            j["metadata"]["name"] != job.metadata.name
+        ), "Should have mutated job name"
+
+    def test_generate_with_template_args(self):
+        mock_config_source = Mock()
+        mock_config_source.get.return_value = V1Job(
+            metadata=V1ObjectMeta(name="anotherone")
+        )
+        generator = JobGenerator(mock_config_source)
+        template_args = {"foo": "bar"}
+
+        generator.generate(template_args=template_args)
+
+        mock_config_source.get.assert_called_once_with(template_args=template_args)
